refactor: 다소 의미가 다른 네이밍 수정
