+++ src/pages/AddCard/domain/validation.ts
