+++ src/stories/AddCardOwnerInput.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import OwnerInput from '../pages/AddCard/components/OwnerInput';
 import useInput from '../hooks/useInput';
 import { APP_WIDTH } from '../utils/constants';
-import { isValidOwnerName } from '../pages/AddCard/domain/dispatcher';
+import { isValidOwnerName } from '../pages/AddCard/domain/validation';
 
 export default {
   title: 'AddCardOwnerInput',
